+++ src/clients/sidecar/apis/index.ts
@@ -9,5 +9,6 @@ export * from "./KafkaConsumeResourceApi";
 export * from "./LoginRealmResourceApi";
 export * from "./MicroProfileHealthApi";
 export * from "./OAuthCallbackResourceApi";
+export * from "./PasswordResetCallbackResourceApi";
 export * from "./PreferencesResourceApi";
 export * from "./VersionResourceApi";